+++ src/backend/backend.h
@@ -50,6 +50,7 @@ public:
     Q_INVOKABLE void removeRequested(const QString id);
     Q_INVOKABLE void pauseAll();
     Q_INVOKABLE void resumeAll();
+    Q_INVOKABLE void removeCompleted();
     Q_INVOKABLE QRect availableScreenGeometry() const;
     Q_INVOKABLE QString coloredSvg(const QString &path, const QString &color);
     Q_INVOKABLE void setCategory(int category);
@@ -80,6 +81,7 @@ private:
     // Functions
     void loadDownloads();
     void wireDownloadConnections(Downloader *downloader, const downloadInformations &info);
+    void removeDownload(const QString &id);
 
     // Variables
     QTcpServer *m_webServer;